+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+* `[sms]` Added network time synchronisation.
 * `[gui][desktop]` Added *how-to-unlock-screen-info* window.
 * `[gui]` Added GUI timers capability
 * `[fonts]` Added multiplication and division sign to the font files
@@ -18,6 +19,7 @@
 
 * `[system]` Timer API - linked to timer, same for Services and Applications. Updated docs
 * `[system]` Removed `using std` and `using cpp_freertos` from commonly used headers
+* `[messages]` Refactored messages SMS thread window to use ListView. 
 
 ### Fixed